requirements.txt
dnspython==1.16.0
elastic-apm==4.2.2
eventlet==0.24.1
Flask==0.11.1
Flask-LogConfig==0.4.2
Flask-Script==2.0.5
greenlet==0.4.13
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==3.0.0 
logconfig==0.4.0
logutils==0.3.5
monotonic==1.5
py==1.5.4
pytest==2.9.2
pytest-cov==2.3.1
PyYAML==3.13
requests==2.10.0
six==1.13.0
urllib3==1.25.10
Werkzeug==0.14.1
pyjwt
python-jose[cryptography]